XMED analysys

+++ Test/BlingoEngine.IO.Legacy.Tests/Texts/BlXmedTestDumper.cs
@@ -41,8 +41,8 @@ namespace BlingoEngine.IO.Legacy.Tests.Texts
                             if (textItem.Format == BlLegacyTextFormatKind.Xmed)
                             {
                                 var path = TestContextHarness.GetAssetPath(folder + "/" + Path.GetFileNameWithoutExtension(item) + "_" + textItem.ResourceId + ".xmed.txt");
-                                //if (File.Exists(path))
-                                //  continue;
+                                if (File.Exists(path))
+                                  continue;
 
                                 File.WriteAllText(path, textItem.Bytes.ToHexString());
                                 var pathBin = TestContextHarness.GetAssetPath(folder + "/" + Path.GetFileNameWithoutExtension(item) + "_" + textItem.ResourceId + ".xmed.bin");
